+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.0].define(version: 2025_07_20_150828) do
+ActiveRecord::Schema[8.0].define(version: 2025_07_27_163749) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
   enable_extension "pgcrypto"
@@ -300,6 +300,8 @@ ActiveRecord::Schema[8.0].define(version: 2025_07_20_150828) do
     t.string "username"
     t.boolean "admin", default: false, null: false
     t.integer "color_schema"
+    t.datetime "discarded_at"
+    t.index ["discarded_at"], name: "index_users_on_discarded_at"
     t.index ["username"], name: "index_users_on_username", unique: true, where: "(username IS NOT NULL)"
   end