+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - changing password
+- discarding users
 
 ## [0.2.8] - 2025-07-27
 ### Added